requirements.txt
Flask==2.3.1
Flask-SQLAlchemy==3.0.3
pandas==2.0.1
plotly==5.14.1
python-dotenv==1.0.0
SQLAlchemy==2.0.10
psycopg2-binary==2.9.6